Move bulk of derive(Bundle) into its own module
